Move HasChainIndex class from cardano-server
